requirements.txt
alembic==1.6.5
amqp==5.0.6
apispec==4.4.2
attrs==21.2.0
bcrypt==3.2.0
billiard==3.6.4.0
blinker==1.4
cached-property==1.5.2
celery==5.1.0
cffi==1.14.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Flask==2.0.1
flask-apispec==0.11.0
Flask-JWT-Extended==4.2.1
Flask-Mail==0.9.1
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
importlib-metadata==4.4.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
kombu==5.1.0
Mako==1.1.4
MarkupSafe==2.0.1
marshmallow==3.12.1
mysqlclient==2.0.3
packaging==20.9
passlib==1.7.4
pluggy==0.13.1
prompt-toolkit==3.0.18
py==1.10.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.17
toml==0.10.2
typing-extensions==3.10.0.0
vine==5.0.0
wcwidth==0.2.5
webargs==8.0.0
Werkzeug==2.0.1
zipp==3.4.1
uwsgi
